Create a python program using Inheritance and perform  pay of workers
python application using Inheritance as per the following. Create a class Worker and
derive two classes DailyWorker and SalariedWorker from it. Every worker has name, salary rate.
Provide a method ComPay(int hours) to compute the week pay of every worker. A DailyWorker is
paid on the basis of number of days he/she works. The SalariedWorker gets paid the wage for 40
hours a week no matter what actual hours is. Implement this scenario to calculate the pay of worker